+++ babel.config.js
@@ -1,6 +1,6 @@
 module.exports = {
-	presets: [
-		["@babel/preset-env", { targets: { node: "current" } }],
-		"@babel/preset-typescript",
-	],
+  presets: [
+    ["@babel/preset-env", { targets: { node: "current" } }],
+    "@babel/preset-typescript",
+  ],
 };
