+++ postcss.config.js
@@ -1,3 +1,3 @@
 module.exports = {
-	plugins: [["autoprefixer"]],
+  plugins: [["autoprefixer"]],
 };
