+++ src/types.ts
@@ -1,0 +1,16 @@
+export interface Task {
+  id: string;
+  text: string;
+  date: string;
+  status: "new" | "in progress" | "done";
+  tags: string[];
+  isDeletedLocally: boolean;
+  isDeletedFromFirebase: boolean;
+}
+
+export interface TaskFilter {
+  text?: string;
+  date?: string;
+  status?: "new" | "in progress" | "done";
+  tags?: string[];
+}
